07-29-2024
Added the current follower 1 control back to the block diagrams. It is needed to convert the analog voltage into a current signal. It must match with the sensitivity setting sent to the Gamry via the USB communication. Next questions: Can you change the voltage in the middle of a watch current experiment? Is there a function I need to send via LV to start reading the voltage from the I monitor port?